[TNT-115] move: InputState model로 옮기기
