+++ feature/trainee/connect/src/main/java/co/kr/tnt/trainee/connect/model/InputState.kt
@@ -1,0 +1,8 @@
+package co.kr.tnt.trainee.connect.model
+
+enum class InputState {
+    UNFOCUSED,
+    FOCUS,
+    VALID,
+    INVALID,
+}
